+++ src/graphql/entities/S3.entity.ts
