+++ src/graphql/entities/document.entity.ts
