+++ src/graphql/entities/permissions.entity.ts
