+++ src/graphql/entities/user.entity.ts
