+++ src/graphql/graphql.module.ts
@@ -4,12 +4,12 @@ import { APP_GUARD } from '@nestjs/core';
 import { GraphQLModule as NestJsGraphQLModule } from '@nestjs/graphql';
 import { resolve } from 'path';
 
-import { S3Service } from '@/_s3/s3.service';
+import { S3Service } from '@/modules/s3/s3.service';
 import { DocumentsService } from '@/modules/documents/documents.service';
 import { FormsService } from '@/modules/forms/forms.service';
 import { UsersService } from '@/modules/users/users.service';
 
-import { DatabaseModule } from '../_database/database.module';
+import { DatabaseModule } from '../modules/database/database.module';
 import { AuthorizationGuard } from './authorization.guard';
 import { DocumentsResolver } from './resolvers/documents.resolver';
 import { FormsResolver } from './resolvers/forms.resolver';
