+++ src/graphql/resolvers/documents.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, Query, Resolver } from '@nestjs/graphql';
 
-import { Roles } from '@/_auth/roles.decorator';
+import { Roles } from '@/modules/auth/roles.decorator';
 import { Document, DocumentsService, ResidueType } from '@/modules/documents';
 import { Role } from '@/shared/utils/constants';
