+++ src/graphql/resolvers/forms.resolver.ts
@@ -7,8 +7,8 @@ import {
   Resolver,
 } from '@nestjs/graphql';
 
-import { AuthUser, CurrentUser } from '@/_auth/current-user';
-import { Roles } from '@/_auth/roles.decorator';
+import { AuthUser, CurrentUser } from '@/modules/auth/current-user';
+import { Roles } from '@/modules/auth/roles.decorator';
 import { DocumentsService } from '@/modules/documents/documents.service';
 import { FormsService } from '@/modules/forms/forms.service';
 import { UsersService } from '@/modules/users/users.service';
