+++ src/graphql/resolvers/users.resolver.ts
@@ -6,12 +6,12 @@ import {
   ResolveField,
   Resolver,
 } from '@nestjs/graphql';
-import { Role } from '@/shared/utils/constants';
 
-import { AuthUser, CurrentUser } from '@/_auth/current-user';
-import { Roles } from '@/_auth/roles.decorator';
+import { AuthUser, CurrentUser } from '@/modules/auth/current-user';
+import { Roles } from '@/modules/auth/roles.decorator';
 import { FormsService } from '@/modules/forms/forms.service';
 import { UsersService } from '@/modules/users/users.service';
+import { Role } from '@/shared/utils/constants';
 
 import { Form } from '../entities/form.entity';
 import { User } from '../entities/user.entity';
