+++ src/graphql/responses/aggregate-form-by-user-profile-response.ts
