+++ src/modules/forms/dtos/find-form.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty, OmitType, PartialType } from '@nestjs/swagger';
 import { IsBoolean, IsOptional } from 'class-validator';
 
-import { PaginationQuery } from '@/shared/dto/paginated.dto';
+import { PaginationQuery } from '@/shared/dtos/paginated.dto';
 import { ToBoolean } from '@/shared/utils/to-boolean';
 
 import { Form } from './form.dto';
