+++ src/modules/forms/dtos/form.dto.ts
@@ -12,7 +12,7 @@ import {
 
 import { Document } from '@/modules/documents';
 import { User } from '@/modules/users';
-import { Timestamp } from '@/shared/dto/timestamp.dto';
+import { Timestamp } from '@/shared/dtos/timestamp.dto';
 import { ToBoolean } from '@/shared/utils/to-boolean';
 
 @ObjectType()
