+++ src/modules/mail/interfaces/mail.interface.ts
