+++ src/modules/users/dtos/find-user.dto.ts
@@ -1,6 +1,6 @@
 import { OmitType, PartialType } from '@nestjs/swagger';
 
-import { PaginationQuery } from '@/shared/dto/paginated.dto';
+import { PaginationQuery } from '@/shared/dtos/paginated.dto';
 
 import { User } from './user.dto';
