+++ src/shared/dtos/s3.dto.ts
